refactor(tests): split tests to multiple files
